TextFile Now being recogniesed, Updated UI so Data will be shown in ComboxFirst the will split the data into Designated Textboxes.
Added in Save Data Button to the UI and code to go along with it, it asks the user if they would like to save the data before doing so.

Need to figure out how to get data into my textBoxs, then finish the gerneration of Random numbers for the Header section.